scripts/requirements_uci.txt
procyclingstats>=0.8.0
gspread>=6.0.0
google-auth>=2.0.0
supabase>=2.0.0
